Hospital_management_system/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.0.335
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langgraph>=0.0.25
python-dotenv>=1.0.0
pydantic>=2.4.2
chromadb>=0.4.18
sentence-transformers>=2.2.2
IPython>=8.15.0
mcp>=0.0.9
langchain-mcp-adapters>=0.0.2
aiosqlite>=0.19.0
orjson>=3.9.10
sqlite-vec>=0.0.1
huggingface-hub>=0.19.4
tavily-python>=0.2.2
requests>=2.31.0